clean `unified.common`
