+++ dlrover/python/unified/backend/elastic/manager.py
@@ -20,7 +20,7 @@ from dlrover.python.unified.backend.elastic import remote_call
 from dlrover.python.unified.backend.elastic.node_check_manager import (
     NodeCheckManager,
 )
-from dlrover.python.unified.common.workload_base import ActorInfo, WorkerStage
+from dlrover.python.unified.common.actor_base import ActorInfo, WorkerStage
 from dlrover.python.unified.common.workload_desc import ElasticWorkloadDesc
 from dlrover.python.unified.controller.api import PrimeMasterApi
 from dlrover.python.unified.util.actor_proxy import (
