+++ dlrover/python/unified/backend/elastic/worker/worker.py
@@ -19,7 +19,7 @@ import torch
 
 from dlrover.python.common.log import default_logger as logger
 from dlrover.python.unified.backend.common.base_worker import BaseWorker
-from dlrover.python.unified.common.workload_base import WorkerStage
+from dlrover.python.unified.common.actor_base import WorkerStage
 from dlrover.python.util.common_util import (
     find_free_port_from_env_and_bind,
 )
