+++ dlrover/python/unified/backend/rl/__init__.py
@@ -10,3 +10,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""This package is deprecated, currently only referenced by veRL example."""
+# TODO remove this package
