+++ dlrover/python/unified/common/constant.py
@@ -12,23 +12,7 @@
 # limitations under the License.
 
 
-class DLMasterConstant(object):
-    JOB_CONTEXT_STATE_KEY = "job-context"
-    SCHEDULING_TIMEOUT_MIN_SECS = 60
-    SCHEDULING_TIMEOUT_PER_ACTOR_SECS = 2
-    SETUP_TIMEOUT_MIN_SECS = 10
-    SETUP_TIMEOUT_PER_ACTOR_SECS = 1
-    RUN_WAIT_INTERVAL = 10
-    EXIT_WAIT_INTERVAL = 5
-    GLOBAL_FAILOVER_INTERVAL = 5
-    WORKLOAD_MAX_RESTART = 30
-    PG_STRATEGY_ENV = "PG_STRATEGY"
-
-
-class DLJobExitReason(object):
-    FINISHED = "FINISHED"
-    ERROR = "ERROR"
-    FAILOVER_OUT_OF_LIMIT = "FAILOVER_OUT_OF_LIMIT"
+RAY_HANG_CHECK_INTERVAL = 5.0  # Interval for monitoring actor invocations
 
 
 class InternalDLWorkloadRole(object):
@@ -50,16 +34,6 @@ class DLWorkloadEnv(object):
 
     DEVICE_COLLOCATION_GROUP = "DEVICE_COLLOCATION_GROUP"
 
-    RAY_SET_VISIBLE_DEVICES_ENVS = {
-        "RAY_EXPERIMENTAL_NOSET_CUDA_VISIBLE_DEVICES": "false",
-        "RAY_EXPERIMENTAL_NOSET_ROCR_VISIBLE_DEVICES": "false",
-        "RAY_EXPERIMENTAL_NOSET_ASCEND_RT_VISIBLE_DEVICES": "false",
-        "RAY_EXPERIMENTAL_NOSET_HABANA_VISIBLE_MODULES": "false",
-        "RAY_EXPERIMENTAL_NOSET_NEURON_RT_VISIBLE_CORES": "false",
-        "RAY_EXPERIMENTAL_NOSET_TPU_VISIBLE_CHIPS": "false",
-        "RAY_EXPERIMENTAL_NOSET_ONEAPI_DEVICE_SELECTOR": "false",
-    }
-
     RAY_NOSET_VISIBLE_DEVICES_ENVS = {
         "RAY_EXPERIMENTAL_NOSET_CUDA_VISIBLE_DEVICES": "true",
         "RAY_EXPERIMENTAL_NOSET_ROCR_VISIBLE_DEVICES": "true",
@@ -69,9 +43,3 @@ class DLWorkloadEnv(object):
         "RAY_EXPERIMENTAL_NOSET_TPU_VISIBLE_CHIPS": "true",
         "RAY_EXPERIMENTAL_NOSET_ONEAPI_DEVICE_SELECTOR": "true",
     }
-
-
-class DLTrainerConstant(object):
-    DEVICE_TYPE_DEFAULT = "GPU"
-    DEVICE_PER_NODE_DEFAULT = 8
-    TORCH_MASTER_PORT_DEFAULT = [21111, 22222, 23333, 24444, 25555]
