+++ dlrover/python/unified/controller/manager.py
@@ -15,19 +15,17 @@ import asyncio
 from typing import List
 
 from dlrover.python.common.log import default_logger as logger
-from dlrover.python.unified.common.workload_base import (
-    MasterStage,
-    WorkerStage,
-)
+from dlrover.python.unified.common.actor_base import WorkerStage
+from dlrover.python.unified.common.enums import MasterStage
 from dlrover.python.unified.util.actor_helper import (
     kill_actors,
     restart_actors,
 )
 from dlrover.python.unified.util.actor_proxy import invoke_actors_t
 
+from ..common.config import JobConfig
 from . import remote_call
 from .api import MasterStatus
-from .config import JobConfig
 from .schedule.graph import DLExecutionGraph
 from .schedule.scheduler import Scheduler
 from .sync_manager import SyncManager
