--- dlrover/python/unified/tests/common/test_exception.py
@@ -1,33 +1,0 @@
-# Copyright 2025 The DLRover Authors. All rights reserved.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from dlrover.python.unified.common.exception import (
-    InvalidDLConfiguration,
-    InvalidJobConfiguration,
-)
-from dlrover.python.unified.tests.base import BaseTest
-
-
-class ExceptionTest(BaseTest):
-    def test_basic(self):
-        with self.assertRaises(InvalidJobConfiguration) as e:
-            raise InvalidJobConfiguration()
-
-        self.assertEqual(str(e.exception), "The job configuration is invalid.")
-
-        with self.assertRaises(InvalidDLConfiguration) as e:
-            raise InvalidDLConfiguration()
-
-        self.assertEqual(
-            str(e.exception), "The deep learning configuration is invalid."
-        )
