+++ dlrover/python/unified/tests/driver/test_driver.py
@@ -15,9 +15,9 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from dlrover.python.unified.common.workload_base import MasterStage
+from dlrover.python.unified.common.config import JobConfig
+from dlrover.python.unified.common.enums import MasterStage
 from dlrover.python.unified.controller.api import PrimeMasterApi
-from dlrover.python.unified.controller.config import JobConfig
 from dlrover.python.unified.driver.main import main, submit
 from dlrover.python.unified.tests.fixtures.example_jobs import (
     elastic_training_job,
